backend/myproject/requirements.txt
django-cors-headers
django
djangorestframework
matplotlib
django-filter